orm+jpa add comment service interface
